Clarify what the outline test module is for
